Include action and action_sub in default metrics
With the following three goals:

1. identify popular and slow actions
2. track performance improvements / regressions on those actions
3. determine if some actions are totally unused in a period of time

One down side to this is sites with a very large number of actions
and subs will have quite a large number of exported metrics. One
way to mitigate this would be make these additional metrics
optional, another is to punt this implementation in to a separate
module or user implementation.